Remove `kube::Service`
Allow creating `kube::Client` from `Service<http::Request<hyper::Body>>` with
some constraints.
